Added to README, really just needed a change of any sort
Github removed the messaging system, not sure how else
to get a hold of the maintainer of this package. As such
this pull request is to send a message to the maintainer.

Nuget package appears to be broken and needs an update:

On Nuget Spatial4n.Core.NTS has a dependency of
NetTopologySuite >= 1.12.0

However, if I try to use the assembly with NTS 1.13.2.0
(current Nuget package) a version mismatch exception occurs

Compiling Spatial4n from source and using that assembly
the exception does not happen. It appears there is a hard
coded version requirement in the Nuget version.